add handle for robot size + virtual obstacles
